Organize code generating functions
